+++ lib/shared/models/ranking_user.freezed.dart
@@ -15,7 +15,7 @@ T _$identity<T>(T value) => value;
 /// @nodoc
 mixin _$RankingUser {
 
- int get rank; String get id; String get fullName; String? get avatarUrl; int get totalPoints; Medal get medal;
+ int get rank; String get id; String get fullName; String? get avatarUrl; String? get avatarSlug; int get totalPoints; Medal get medal;
 /// Create a copy of RankingUser
 /// with the given fields replaced by the non-null parameter values.
 @JsonKey(includeFromJson: false, includeToJson: false)
@@ -28,16 +28,16 @@ $RankingUserCopyWith<RankingUser> get copyWith => _$RankingUserCopyWithImpl<Rank
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is RankingUser&&(identical(other.rank, rank) || other.rank == rank)&&(identical(other.id, id) || other.id == id)&&(identical(other.fullName, fullName) || other.fullName == fullName)&&(identical(other.avatarUrl, avatarUrl) || other.avatarUrl == avatarUrl)&&(identical(other.totalPoints, totalPoints) || other.totalPoints == totalPoints)&&(identical(other.medal, medal) || other.medal == medal));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is RankingUser&&(identical(other.rank, rank) || other.rank == rank)&&(identical(other.id, id) || other.id == id)&&(identical(other.fullName, fullName) || other.fullName == fullName)&&(identical(other.avatarUrl, avatarUrl) || other.avatarUrl == avatarUrl)&&(identical(other.avatarSlug, avatarSlug) || other.avatarSlug == avatarSlug)&&(identical(other.totalPoints, totalPoints) || other.totalPoints == totalPoints)&&(identical(other.medal, medal) || other.medal == medal));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,rank,id,fullName,avatarUrl,totalPoints,medal);
+int get hashCode => Object.hash(runtimeType,rank,id,fullName,avatarUrl,avatarSlug,totalPoints,medal);
 
 @override
 String toString() {
-  return 'RankingUser(rank: $rank, id: $id, fullName: $fullName, avatarUrl: $avatarUrl, totalPoints: $totalPoints, medal: $medal)';
+  return 'RankingUser(rank: $rank, id: $id, fullName: $fullName, avatarUrl: $avatarUrl, avatarSlug: $avatarSlug, totalPoints: $totalPoints, medal: $medal)';
 }
 
 
@@ -48,7 +48,7 @@ abstract mixin class $RankingUserCopyWith<$Res>  {
   factory $RankingUserCopyWith(RankingUser value, $Res Function(RankingUser) _then) = _$RankingUserCopyWithImpl;
 @useResult
 $Res call({
- int rank, String id, String fullName, String? avatarUrl, int totalPoints, Medal medal
+ int rank, String id, String fullName, String? avatarUrl, String? avatarSlug, int totalPoints, Medal medal
 });
 
 
@@ -65,12 +65,13 @@ class _$RankingUserCopyWithImpl<$Res>
 
 /// Create a copy of RankingUser
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? rank = null,Object? id = null,Object? fullName = null,Object? avatarUrl = freezed,Object? totalPoints = null,Object? medal = null,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? rank = null,Object? id = null,Object? fullName = null,Object? avatarUrl = freezed,Object? avatarSlug = freezed,Object? totalPoints = null,Object? medal = null,}) {
   return _then(_self.copyWith(
 rank: null == rank ? _self.rank : rank // ignore: cast_nullable_to_non_nullable
 as int,id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
 as String,fullName: null == fullName ? _self.fullName : fullName // ignore: cast_nullable_to_non_nullable
 as String,avatarUrl: freezed == avatarUrl ? _self.avatarUrl : avatarUrl // ignore: cast_nullable_to_non_nullable
+as String?,avatarSlug: freezed == avatarSlug ? _self.avatarSlug : avatarSlug // ignore: cast_nullable_to_non_nullable
 as String?,totalPoints: null == totalPoints ? _self.totalPoints : totalPoints // ignore: cast_nullable_to_non_nullable
 as int,medal: null == medal ? _self.medal : medal // ignore: cast_nullable_to_non_nullable
 as Medal,
@@ -158,10 +159,10 @@ return $default(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( int rank,  String id,  String fullName,  String? avatarUrl,  int totalPoints,  Medal medal)?  $default,{required TResult orElse(),}) {final _that = this;
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( int rank,  String id,  String fullName,  String? avatarUrl,  String? avatarSlug,  int totalPoints,  Medal medal)?  $default,{required TResult orElse(),}) {final _that = this;
 switch (_that) {
 case _RankingUser() when $default != null:
-return $default(_that.rank,_that.id,_that.fullName,_that.avatarUrl,_that.totalPoints,_that.medal);case _:
+return $default(_that.rank,_that.id,_that.fullName,_that.avatarUrl,_that.avatarSlug,_that.totalPoints,_that.medal);case _:
   return orElse();
 
 }
@@ -179,10 +180,10 @@ return $default(_that.rank,_that.id,_that.fullName,_that.avatarUrl,_that.totalPo
 /// }
 /// ```
 
-@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( int rank,  String id,  String fullName,  String? avatarUrl,  int totalPoints,  Medal medal)  $default,) {final _that = this;
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( int rank,  String id,  String fullName,  String? avatarUrl,  String? avatarSlug,  int totalPoints,  Medal medal)  $default,) {final _that = this;
 switch (_that) {
 case _RankingUser():
-return $default(_that.rank,_that.id,_that.fullName,_that.avatarUrl,_that.totalPoints,_that.medal);case _:
+return $default(_that.rank,_that.id,_that.fullName,_that.avatarUrl,_that.avatarSlug,_that.totalPoints,_that.medal);case _:
   throw StateError('Unexpected subclass');
 
 }
@@ -199,10 +200,10 @@ return $default(_that.rank,_that.id,_that.fullName,_that.avatarUrl,_that.totalPo
 /// }
 /// ```
 
-@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( int rank,  String id,  String fullName,  String? avatarUrl,  int totalPoints,  Medal medal)?  $default,) {final _that = this;
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( int rank,  String id,  String fullName,  String? avatarUrl,  String? avatarSlug,  int totalPoints,  Medal medal)?  $default,) {final _that = this;
 switch (_that) {
 case _RankingUser() when $default != null:
-return $default(_that.rank,_that.id,_that.fullName,_that.avatarUrl,_that.totalPoints,_that.medal);case _:
+return $default(_that.rank,_that.id,_that.fullName,_that.avatarUrl,_that.avatarSlug,_that.totalPoints,_that.medal);case _:
   return null;
 
 }
@@ -214,13 +215,14 @@ return $default(_that.rank,_that.id,_that.fullName,_that.avatarUrl,_that.totalPo
 @JsonSerializable()
 
 class _RankingUser implements RankingUser {
-  const _RankingUser({required this.rank, required this.id, required this.fullName, this.avatarUrl, required this.totalPoints, this.medal = Medal.hierro});
+  const _RankingUser({required this.rank, required this.id, required this.fullName, this.avatarUrl, this.avatarSlug, required this.totalPoints, this.medal = Medal.hierro});
   factory _RankingUser.fromJson(Map<String, dynamic> json) => _$RankingUserFromJson(json);
 
 @override final  int rank;
 @override final  String id;
 @override final  String fullName;
 @override final  String? avatarUrl;
+@override final  String? avatarSlug;
 @override final  int totalPoints;
 @override@JsonKey() final  Medal medal;
 
@@ -237,16 +239,16 @@ Map<String, dynamic> toJson() {
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _RankingUser&&(identical(other.rank, rank) || other.rank == rank)&&(identical(other.id, id) || other.id == id)&&(identical(other.fullName, fullName) || other.fullName == fullName)&&(identical(other.avatarUrl, avatarUrl) || other.avatarUrl == avatarUrl)&&(identical(other.totalPoints, totalPoints) || other.totalPoints == totalPoints)&&(identical(other.medal, medal) || other.medal == medal));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _RankingUser&&(identical(other.rank, rank) || other.rank == rank)&&(identical(other.id, id) || other.id == id)&&(identical(other.fullName, fullName) || other.fullName == fullName)&&(identical(other.avatarUrl, avatarUrl) || other.avatarUrl == avatarUrl)&&(identical(other.avatarSlug, avatarSlug) || other.avatarSlug == avatarSlug)&&(identical(other.totalPoints, totalPoints) || other.totalPoints == totalPoints)&&(identical(other.medal, medal) || other.medal == medal));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,rank,id,fullName,avatarUrl,totalPoints,medal);
+int get hashCode => Object.hash(runtimeType,rank,id,fullName,avatarUrl,avatarSlug,totalPoints,medal);
 
 @override
 String toString() {
-  return 'RankingUser(rank: $rank, id: $id, fullName: $fullName, avatarUrl: $avatarUrl, totalPoints: $totalPoints, medal: $medal)';
+  return 'RankingUser(rank: $rank, id: $id, fullName: $fullName, avatarUrl: $avatarUrl, avatarSlug: $avatarSlug, totalPoints: $totalPoints, medal: $medal)';
 }
 
 
@@ -257,7 +259,7 @@ abstract mixin class _$RankingUserCopyWith<$Res> implements $RankingUserCopyWith
   factory _$RankingUserCopyWith(_RankingUser value, $Res Function(_RankingUser) _then) = __$RankingUserCopyWithImpl;
 @override @useResult
 $Res call({
- int rank, String id, String fullName, String? avatarUrl, int totalPoints, Medal medal
+ int rank, String id, String fullName, String? avatarUrl, String? avatarSlug, int totalPoints, Medal medal
 });
 
 
@@ -274,12 +276,13 @@ class __$RankingUserCopyWithImpl<$Res>
 
 /// Create a copy of RankingUser
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? rank = null,Object? id = null,Object? fullName = null,Object? avatarUrl = freezed,Object? totalPoints = null,Object? medal = null,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? rank = null,Object? id = null,Object? fullName = null,Object? avatarUrl = freezed,Object? avatarSlug = freezed,Object? totalPoints = null,Object? medal = null,}) {
   return _then(_RankingUser(
 rank: null == rank ? _self.rank : rank // ignore: cast_nullable_to_non_nullable
 as int,id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
 as String,fullName: null == fullName ? _self.fullName : fullName // ignore: cast_nullable_to_non_nullable
 as String,avatarUrl: freezed == avatarUrl ? _self.avatarUrl : avatarUrl // ignore: cast_nullable_to_non_nullable
+as String?,avatarSlug: freezed == avatarSlug ? _self.avatarSlug : avatarSlug // ignore: cast_nullable_to_non_nullable
 as String?,totalPoints: null == totalPoints ? _self.totalPoints : totalPoints // ignore: cast_nullable_to_non_nullable
 as int,medal: null == medal ? _self.medal : medal // ignore: cast_nullable_to_non_nullable
 as Medal,